requirements.txt
flask
flask-cors
pymongo
tensorflow
keras 
opencv-python 
scipy
numpy==1.23.5
tensorflow==2.14
keras==2.14 
h5py
pymongo
pillow
numpy
deepface
werkzeug
python-dotenv
gunicorn
twilio
